Refresh week's classes daily and today's bi-hourly
We don't need to warm the cache with the week's data multiple times in a
day. We assume that it is not important to refresh classes later in the
week more than once a day (for subs).

Also throw exception if CHROME_WEBDRIVER_PATH env is not set. Currently
this silently fails.